+++ src/props/mod.rs
@@ -26,7 +26,9 @@
  * SOFTWARE.
  */
 // ext
+use std::collections::{HashMap, LinkedList};
 use tui::style::{Color, Modifier, Style};
+use tui::widgets::canvas::{Line, Map, Rectangle};
 
 // modules
 pub mod borders;
@@ -39,8 +41,6 @@ pub use borders::{Borders, BordersProps};
 pub use builder::{GenericPropsBuilder, PropsBuilder};
 pub use dataset::Dataset;
 pub use texts::{Table, TableBuilder, TextParts, TextSpan, TextSpanBuilder};
-
-use std::collections::{HashMap, LinkedList};
 
 // -- Props
 
@@ -117,6 +117,7 @@ pub enum PropValue {
     Dataset(Dataset),
     Color(Color),
     InputType(InputType),
+    Shape(Shape),
     Style(Style),
 }
 
@@ -130,6 +131,43 @@ pub enum InputType {
     Text,
     Number,
     Password,
+}
+
+// -- Shape
+
+/// ## Shape
+///
+/// Describes the shape to draw on the canvas
+#[derive(Clone, Debug)]
+pub enum Shape {
+    //Label((f64, f64, String, Color)),
+    Layer,
+    Line(Line),
+    Map(Map),
+    Points((Vec<(f64, f64)>, Color)),
+    Rectangle(Rectangle),
+}
+
+impl PartialEq for Shape {
+    fn eq(&self, other: &Self) -> bool {
+        match (self, other) {
+            //(Shape::Label(a), Shape::Label(b)) => a == b,
+            (Shape::Layer, Shape::Layer) => true,
+            (Shape::Line(a), Shape::Line(b)) => {
+                a.x1 == b.x1 && a.x2 == b.x2 && a.y1 == b.y1 && a.y2 == b.y2 && a.color == b.color
+            }
+            (Shape::Map(a), Shape::Map(b)) => a.color == b.color,
+            (Shape::Points(a), Shape::Points(b)) => a == b,
+            (Shape::Rectangle(a), Shape::Rectangle(b)) => {
+                a.x == b.x
+                    && a.y == b.y
+                    && a.width == b.width
+                    && a.height == b.height
+                    && a.color == b.color
+            }
+            (_, _) => false,
+        }
+    }
 }
 
 #[cfg(test)]
@@ -229,6 +267,14 @@ mod tests {
         assert_eq!(
             *map.get("s").unwrap(),
             PropValue::InputType(InputType::Number)
+        );
+        map.insert(
+            String::from("t"),
+            PropValue::Shape(Shape::Map(Map::default())),
+        );
+        assert_eq!(
+            *map.get("t").unwrap(),
+            PropValue::Shape(Shape::Map(Map::default()))
         );
         PropPayload::Map(map);
         let mut link: LinkedList<PropPayload> = LinkedList::new();